+++ package.json
@@ -17,6 +17,7 @@
     "billboard-top-100": "^2.6.5",
     "date-fns": "^4.1.0",
     "framer-motion": "^11.17.0",
+    "genius-lyrics": "^4.4.7",
     "next": "14.2.20",
     "next-themes": "^0.4.4",
     "react": "^18",
