+++ src/app/api/lyrics/route.ts
@@ -1,6 +1,7 @@
 import { NextResponse } from 'next/server'
 
-export async function GET() {
+export async function GET(request: Request) {
+  // 1. accessToken 키 발급
   const clientId = process.env.GENIUS_CLIENT_ID
   const clientSecret = process.env.GENIUS_CLIENT_SECRET
 
@@ -37,7 +38,30 @@ export async function GET() {
     const tokenData = await tokenResponse.json()
     const accessToken = tokenData.access_token
 
-    return NextResponse.json({ accessToken }, { status: 200 })
+    // 2. 가사 갖고오기
+    const { searchParams } = new URL(request.url)
+    const artist = searchParams.get('artist')
+    const title = searchParams.get('title')
+
+    if (!artist || !title) {
+      return NextResponse.json(
+        {
+          error:
+            'Missing required query parameters: artist, title, or authToken.',
+        },
+        { status: 400 },
+      )
+    }
+
+    const query = `${artist} ${title}`
+
+    const Genius = require('genius-lyrics')
+    const client = new Genius.Client(accessToken)
+    const searches = await client.songs.search(query)
+    const firstSong = searches[0]
+    const lyrics = await firstSong.lyrics()
+
+    return NextResponse.json({ lyrics }, { status: 200 })
   } catch (error) {
     const errorMessage =
       error instanceof Error ? error.message : 'An unknown error occurred.'
